requirements.txt
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.147.0
jinja2==3.1.4
PyYAML==6.0.2
python-dateutil==2.9.0.post0
pandas==2.2.3
pytest